+++ allure-python-commons/src/_core.py
@@ -5,7 +5,9 @@ from allure_commons import _hooks
 
 _storage = threading.local()
 _storage.plugin_manager = PluginManager('allure')
-_storage.plugin_manager.add_hookspecs(_hooks)
+_storage.plugin_manager.add_hookspecs(_hooks.AllureUserHooks)
+_storage.plugin_manager.add_hookspecs(_hooks.AllureDeveloperHooks)
+
 
 plugin_manager = _storage.plugin_manager
 register = plugin_manager.register
