+++ allure-python-commons/src/_hooks.py
@@ -4,51 +4,63 @@ hookspec = HookspecMarker("allure")
 hookimpl = HookimplMarker("allure")
 
 
-@hookspec
-def decorate_as_label(label_type, labels):
-    """ label """
+class AllureUserHooks(object):
+
+    @hookspec
+    def decorate_as_label(self, label_type, labels):
+        """ label """
+
+    @hookspec
+    def add_label(self, label_type, labels):
+        """ label """
+
+    @hookspec
+    def decorate_as_link(self, url, link_type, name):
+        """ url """
+
+    @hookspec
+    def add_link(self, url, link_type, name):
+        """ url """
+
+    @hookspec
+    def start_step(self, uuid, title, params):
+        """ step """
+
+    @hookspec
+    def stop_step(self, uuid, exc_type, exc_val, exc_tb):
+        """ step """
+
+    @hookspec
+    def attach_data(self, body, name, attachment_type, extension):
+        """ attach data """
+
+    @hookspec
+    def attach_file(self, source, name, attachment_type, extension):
+        """ attach file """
 
 
-@hookspec
-def add_label(label_type, labels):
-    """ label """
+class AllureDeveloperHooks(object):
 
+    @hookspec
+    def start_fixture(self, parent_uuid, uuid, name):
+        """ start fixture"""
 
-@hookspec
-def decorate_as_link(url, link_type, name):
-    """ url """
+    @hookspec
+    def stop_fixture(self, uuid, exc_type, exc_val, exc_tb):
+        """ stop fixture """
 
+    @hookspec
+    def report_result(self, result):
+        """ reporting """
 
-@hookspec
-def add_link(url, link_type, name):
-    """ url """
+    @hookspec
+    def report_container(self, container):
+        """ reporting """
 
+    @hookspec
+    def report_attached_file(self, source, file_name):
+        """ reporting """
 
-@hookspec
-def start_step(uuid, title, params):
-    """ step """
-
-
-@hookspec
-def stop_step(uuid, exc_type, exc_val, exc_tb):
-    """ step """
-
-
-@hookspec
-def attach_data(body, name, attachment_type, extension):
-    """ attach data """
-
-
-@hookspec
-def attach_file(source, name, attachment_type, extension):
-    """ attach file """
-
-
-@hookspec
-def start_fixture(parent_uuid, uuid, name):
-    """ start fixture"""
-
-
-@hookspec
-def stop_fixture(uuid, exc_type, exc_val, exc_tb):
-    """ stop fixture """
+    @hookspec
+    def report_attached_data(self, body, file_name):
+        """ reporting """
